Ep13: 게시물 저자 본인만 삭제할 수 있도록 구현

+++ backend/api/resources/post.py
@@ -49,10 +49,20 @@ class Post(Resource):
     @classmethod
     @jwt_required()
     def delete(cls, id):
+        # 요청을 보낸 사용자
+        username = get_jwt_identity()
+        author_id = UserModel.find_by_username(username).id
+
         post = PostModel.find_by_id(id)
+        
+        # 게시물이 존재하는지 확인
         if post:
-            post.delete_from_db()
-            return {"message": "게시물이 성공적으로 삭제되었습니다."}, 200
+            # 게시물의 작성자와 요청을 보낸 사용자가 같은지 확인
+            if post.author_id == author_id:
+                post.delete_from_db()
+                return {"message": "게시물이 성공적으로 삭제되었습니다."}, 200
+            else:
+                return {"Error": "게시물은 작성자만 삭제할 수 있습니다."}, 403
         return {"Error": "게시물을 찾을 수 없습니다."}, 404
